telegram_scraper/requirements.txt
telethon==1.34.0
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pydantic==2.5.3
httpx==0.26.0
pillow==10.2.0
google-cloud-storage==2.14.0
python-dateutil==2.8.2
loguru==0.7.2
typer==0.9.0
aiofiles==23.2.1
asyncio-throttle==1.0.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 